refactor: Move `augmentation_config` to `datasets/aug_config` and update imports in codebase and docs
