+++ tests/datasets/test_augmentations.py
@@ -7,13 +7,12 @@
 """Tests for Albumentations augmentation wrappers."""
 
 import albumentations as A
-import numpy as np
 import pytest
 import torch
 from PIL import Image
 from torch.utils.data import DataLoader
 
-from rfdetr.augmentation_config import AUG_CONFIG
+from rfdetr.datasets.aug_config import AUG_CONFIG
 from rfdetr.datasets.transforms import (
     AlbumentationsWrapper,
     Compose,
